Rename manager modules to snake_case
