+++ glacium/cli/job/run.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import click
 
 from glacium.utils.current import load
-from glacium.managers.ProjectManager import ProjectManager
+from glacium.managers.project_manager import ProjectManager
 from glacium.models.job import JobStatus
 
 from . import cli_job, ROOT
