+++ glacium/managers/template_manager.py
