+++ glacium/recipes/fensap.py
@@ -1,6 +1,6 @@
 """Recipe containing jobs to run the FENSAP solver."""
 
-from glacium.managers.RecipeManager import RecipeManager, BaseRecipe
+from glacium.managers.recipe_manager import RecipeManager, BaseRecipe
 from glacium.jobs.fensap_jobs import FensapRunJob
 
 @RecipeManager.register
