+++ glacium/recipes/pointwise.py
@@ -1,6 +1,6 @@
 """Recipe integrating Pointwise mesh generation jobs."""
 
-from glacium.managers.RecipeManager import RecipeManager, BaseRecipe
+from glacium.managers.recipe_manager import RecipeManager, BaseRecipe
 from glacium.jobs.pointwise_jobs import PointwiseGCIJob, PointwiseMesh2Job
 from glacium.engines.fluent2fensap import Fluent2FensapJob
