+++ tests/test_fluent2fensap_default.py
@@ -5,7 +5,7 @@ sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
 from glacium.engines.fluent2fensap import Fluent2FensapJob
 from glacium.engines.base_engine import BaseEngine
 from glacium.models.config import GlobalConfig
-from glacium.managers.PathManager import PathBuilder, _SharedState
+from glacium.managers.path_manager import PathBuilder, _SharedState
 from glacium.models.project import Project
